mv BreadCrumbs and SearchResults to separate components outside Bucket
