+++ catalog/app/components/BreadCrumbs/index.js
@@ -1,0 +1,2 @@
+export * from './BreadCrumbs'
+export default from './BreadCrumbs'
