+++ catalog/app/components/SearchResults/index.js
@@ -1,0 +1,1 @@
+export * from './SearchResults'
