+++ catalog/app/containers/Bucket/Dir.js
@@ -5,6 +5,8 @@ import * as R from 'ramda'
 import * as React from 'react'
 import { Box, CircularProgress } from '@material-ui/core'
 
+import BreadCrumbs, { Crumb } from 'components/BreadCrumbs'
+import { docs } from 'constants/urls'
 import AsyncResult from 'utils/AsyncResult'
 import * as AWS from 'utils/AWS'
 import Data from 'utils/Data'
@@ -12,8 +14,6 @@ import * as NamedRoutes from 'utils/NamedRoutes'
 import Link from 'utils/StyledLink'
 import { getBreadCrumbs, ensureNoSlash, withoutPrefix, up } from 'utils/s3paths'
 
-import { docs } from 'constants/urls'
-import BreadCrumbs, { Crumb } from './BreadCrumbs'
 import Code from './Code'
 import Listing, { ListingItem } from './Listing'
 import Message from './Message'
