+++ catalog/app/containers/Bucket/File.js
@@ -25,6 +25,7 @@ import {
 } from '@material-ui/core'
 import { makeStyles, styled } from '@material-ui/styles'
 
+import BreadCrumbs, { Crumb } from 'components/BreadCrumbs'
 import ButtonIcon from 'components/ButtonIcon'
 import Sparkline from 'components/Sparkline'
 import AsyncResult from 'utils/AsyncResult'
@@ -39,7 +40,6 @@ import * as RT from 'utils/reactTools'
 import { getBreadCrumbs, up } from 'utils/s3paths'
 import { readableBytes, readableQuantity } from 'utils/string'
 
-import BreadCrumbs, { Crumb } from './BreadCrumbs'
 import Code from './Code'
 import FilePreview from './FilePreview'
 import Section from './Section'
