+++ catalog/app/containers/Bucket/PackageTree.js
@@ -6,6 +6,7 @@ import * as React from 'react'
 import { Box, Button, CircularProgress } from '@material-ui/core'
 import { makeStyles } from '@material-ui/styles'
 
+import BreadCrumbs, { Crumb } from 'components/BreadCrumbs'
 import { ThrowNotFound } from 'containers/NotFoundPage'
 import AsyncResult from 'utils/AsyncResult'
 import * as AWS from 'utils/AWS'
@@ -15,7 +16,6 @@ import * as NamedRoutes from 'utils/NamedRoutes'
 import { getBreadCrumbs, getPrefix, isDir, parseS3Url, up } from 'utils/s3paths'
 import tagged from 'utils/tagged'
 
-import BreadCrumbs, { Crumb } from './BreadCrumbs'
 import Code from './Code'
 import FilePreview from './FilePreview'
 import Listing, { ListingItem } from './Listing'
